Pruebas entrega 2 terminada

+++ src/fiuba/algo3/algoFormers/generico/Potencia.java
@@ -1,8 +1,9 @@
 package fiuba.algo3.algoFormers.generico;
 
 public class Potencia {
-	float potencia = 1;
+	double potencia = 1;
 	public Potencia(){}
+	
 	public void multiplicarAtaque(double i){
 		this.potencia*=i;
 	}
